+++ PKHeX.Core/Legality/Encounters/EncounterStatic/EncounterTera9.cs
@@ -132,7 +132,7 @@ public sealed record EncounterTera9 : EncounterStatic, ITeraRaid9
             return true;
 
         var pi = PersonalTable.SV.GetFormEntry(Species, Form);
-        var param = new GenerateParam9((byte)pi.Gender, FlawlessIVCount, 1, 0, 0, 0, Ability, Shiny);
+        var param = new GenerateParam9(pi.Gender, FlawlessIVCount, 1, 0, 0, 0, Ability, Shiny);
         if (!Encounter9RNG.IsMatch(pk, param, seed))
             return true;
         return base.IsMatchPartial(pk);
@@ -152,7 +152,7 @@ public sealed record EncounterTera9 : EncounterStatic, ITeraRaid9
         const byte rollCount = 1;
         const byte undefinedSize = 0;
         var pi = PersonalTable.SV.GetFormEntry(Species, Form);
-        var param = new GenerateParam9((byte)pi.Gender, FlawlessIVCount, rollCount,
+        var param = new GenerateParam9(pi.Gender, FlawlessIVCount, rollCount,
             undefinedSize, undefinedSize, undefinedSize,
             Ability, Shiny);
