+++ PKHeX.Core/MysteryGifts/WR7.cs
@@ -87,8 +87,8 @@ public sealed class WR7 : DataMysteryGift
     public override int Gender { get; set; }
     public override byte Form { get; set; }
     public override uint ID32 { get; set; }
-    public override uint TID16 { get; set; }
-    public override uint SID16 { get; set; }
+    public override ushort TID16 { get; set; }
+    public override ushort SID16 { get; set; }
 
     public override string OT_Name
     {
